Use interface for repository
